refactor: move Network files from src/common/ to src/common/network/
