+++ src/common/network/NetworkCommon.cpp
@@ -1,4 +1,4 @@
-#include "common/NetworkCommon.hpp"
+#include "common/network/NetworkCommon.hpp"
 
 #include <QStringList>
