+++ src/common/network/NetworkManager.hpp
