+++ src/common/network/NetworkPrivate.cpp
@@ -1,8 +1,8 @@
-#include "common/NetworkPrivate.hpp"
+#include "common/network/NetworkPrivate.hpp"
 
 #include "common/network/NetworkTask.hpp"
-#include "common/NetworkManager.hpp"
-#include "common/NetworkResult.hpp"
+#include "common/network/NetworkManager.hpp"
+#include "common/network/NetworkResult.hpp"
 #include "common/QLogging.hpp"
 #include "singletons/Paths.hpp"
 #include "util/AbandonObject.hpp"
