+++ src/common/network/NetworkResult.hpp
