+++ src/providers/chatterino/ChatterinoBadges.cpp
@@ -1,7 +1,7 @@
 #include "ChatterinoBadges.hpp"
 
-#include "common/NetworkRequest.hpp"
-#include "common/NetworkResult.hpp"
+#include "common/network/NetworkRequest.hpp"
+#include "common/network/NetworkResult.hpp"
 #include "messages/Emote.hpp"
 
 #include <QJsonArray>
