+++ src/providers/ffz/FfzBadges.cpp
@@ -1,7 +1,7 @@
 #include "FfzBadges.hpp"
 
-#include "common/NetworkRequest.hpp"
-#include "common/NetworkResult.hpp"
+#include "common/network/NetworkRequest.hpp"
+#include "common/network/NetworkResult.hpp"
 #include "messages/Emote.hpp"
 #include "providers/ffz/FfzUtil.hpp"
