+++ src/providers/recentmessages/Api.cpp
@@ -1,7 +1,7 @@
 #include "providers/recentmessages/Api.hpp"
 
-#include "common/NetworkRequest.hpp"
-#include "common/NetworkResult.hpp"
+#include "common/network/NetworkRequest.hpp"
+#include "common/network/NetworkResult.hpp"
 #include "common/QLogging.hpp"
 #include "providers/recentmessages/Impl.hpp"
 #include "providers/twitch/TwitchMessageBuilder.hpp"
