+++ src/providers/seventv/SeventvAPI.cpp
@@ -1,8 +1,8 @@
 #include "providers/seventv/SeventvAPI.hpp"
 
 #include "common/Literals.hpp"
-#include "common/NetworkRequest.hpp"
-#include "common/NetworkResult.hpp"
+#include "common/network/NetworkRequest.hpp"
+#include "common/network/NetworkResult.hpp"
 
 namespace {
