+++ src/providers/twitch/api/Helix.cpp
@@ -1,8 +1,8 @@
 #include "providers/twitch/api/Helix.hpp"
 
 #include "common/Literals.hpp"
-#include "common/NetworkRequest.hpp"
-#include "common/NetworkResult.hpp"
+#include "common/network/NetworkRequest.hpp"
+#include "common/network/NetworkResult.hpp"
 #include "common/QLogging.hpp"
 #include "util/CancellationToken.hpp"
