+++ tests/src/NetworkCommon.cpp
@@ -1,4 +1,4 @@
-#include "common/NetworkCommon.hpp"
+#include "common/network/NetworkCommon.hpp"
 
 #include <gtest/gtest.h>
